Refactor xds test driver to prepare for additional test cases
The bulk of this change is the introduction of a GcpResource object to
track the created VMs, backend services, URL maps, etc. The additional
test cases for xDS integration greatly increase the complexity of the
resources required (e.g., multiple backend services, instance groups in
different zones) and the GcpResource construct makes keeping track of -
and cleaning up - these resources much cleaner.