Add --server_port flag to python benchmark worker
The benchmark workers for most gRPC languages support a `--server_port`
flag. When specified, the server accepts traffic on this port. This is
true even if a port is specified in the ServerConfig message of a
Scenario.

This change adds this flag to the benchmark worker for Python, allowing
it to adopt the same behavior.