Remove TCP transmit zerocopy shutdown when removing TCP endpoint from pollset. (#25782)
Previously, we disabled TCP transmit zerocopy and waited for all pending
completions when removing a TCP endpoint from a pollset. However, certain
transports may do this and then re-add the endpoint to a pollset, in which case
zerocopy is no longer active.

Instead, here we only shutdown zerocopy if the socket is being torn down.