Correctly reset msg_controllen before processing TCP error queue. (#26250)
When we call recvmsg(), we are allowed to set msg_control and
msg_controllen to receive ancillary messages from the TCP socket.
msg_controllen is set to available buffer length for such sockets
before we call recvmsg(); after the call returns successfully, it
contains the amount of available data in the ancillary buffer.

Existing behaviour is buggy; it calls recvmsg() in a loop but does not
set msg_controllen to the size of the full buffer correctly on each
iteration. This leads to a surfeit of error log messages, claiming the
ancillary messages had to be truncated due to insufficient buffer
space.

This patch correctly sets the ancillary buffer size before each call
to recvmsg() when processing the TCP error queue.