Add experimental CI for PSM tests. (#29477)
This commit adds experimental CI for PSM tests.

Test included initially are Java and cxx tests. With expected load
from 100 to 30000, same as the manual test. Running proxied and proxyless
tests on benchmark-prod2 cluster, 8-core-machines (both client and server).

Data are uploaded to experimental tables.

Open for initial feedback on what we want to run on CI.