Add std::shared_ptr support to ChannelArgs, and precondition ChannelArgs with a default EventEngine (#30128)
* [WIP] Precondition ChannelArgs with a default EventEngine

This is a step towards using ChannelArgs as the primary means of
accessing EventEngine instances in gRPC-core. If not explicitly provided
by the application, a default EventEngine will populated into
ChannelArgs during preconditioning.

This is not a final state, we may want to enable ref-counting here
instead of using raw pointers. And a refactoring is in order to enable
GetObject instead of the more verbose
GetPointer<EventEngine>(GRPC_ARG_EVENT_ENGINE).

* Refactor ChannelArgs::GetObject to support non-conforming classes

This allows us to not expose `ChannelArgName` in the public interface.

* Add std::shared_ptr to ChannelArgs; Add EventEngine specialization

* subchannel fix; cleanup

* replace GetSharedPtr with overloads of GetObjectRef

* Automated change: Fix sanity tests

* fix the fixer

* fix raw pointer retrieval from stored shared_ptr

* Make GetObjectRef<EventEngine> work (not general to shared_ptr)

* enable shared_ptr ChannelArg support for shared_from_this

* use new EventEngines for tests (not the default global)

* Automated change: Fix sanity tests

Co-authored-by: drfloob <drfloob@users.noreply.github.com>